[#1297] moved file to correct package + some formatting
